app/core/knowledge_base.py
"""
Knowledge Base Module - Knowledge Management

TODO: Implement the following features
1. Initialization and management of the travel knowledge base  
2. Knowledge updating and synchronization  
3. Knowledge quality evaluation  
4. Multilingual knowledge support  
5. Knowledge categorization and tag management  
"""

from typing import Dict, List, Optional, Any
from pathlib import Path
import json
import yaml
from pydantic import BaseModel
from app.core.rag_engine import Document, RAGEngine, get_rag_engine


class KnowledgeCategory(BaseModel):
    """知识分类"""
    id: str
    name: str
    description: str
    parent_id: Optional[str] = None
    metadata: Dict[str, Any] = {}


class KnowledgeSource(BaseModel):
    """知识来源"""
    id: str
    name: str
    url: Optional[str] = None
    last_updated: Optional[str] = None
    reliability_score: float = 1.0
    language: str = "zh"


class TravelKnowledge(BaseModel):
    """旅游知识结构"""
    id: str
    title: str
    content: str
    category: str
    location: Optional[str] = None
    tags: List[str] = []
    source: Optional[KnowledgeSource] = None
    language: str = "zh"
    last_updated: Optional[str] = None


class KnowledgeBase:
    """知识库管理器"""
    
    def __init__(self, knowledge_dir: str = "app/knowledge"):
        self.knowledge_dir = Path(knowledge_dir)
        self.categories: Dict[str, KnowledgeCategory] = {}
        self.knowledge_items: Dict[str, TravelKnowledge] = {}
        self.rag_engine = get_rag_engine()
    
    async def initialize(self):
        """初始化知识库"""
        # TODO: 实现知识库初始化
        # 1. 加载知识分类
        # 2. 加载基础知识数据
        # 3. 建立索引
        await self._load_categories()
        await self._load_knowledge_data()
        await self._build_index()
    
    async def _load_categories(self):
        """加载知识分类"""
        # TODO: 从配置文件加载知识分类
        categories_file = self.knowledge_dir / "categories.yaml"
        if categories_file.exists():
            # 加载分类配置
            pass
        else:
            # 创建默认分类
            await self._create_default_categories()
    
    async def _create_default_categories(self):
        """创建默认知识分类"""
        default_categories = [
            {
                "id": "destinations",
                "name": "目的地信息",
                "description": "各地旅游目的地的基本信息、景点、文化等"
            },
            {
                "id": "transportation",
                "name": "交通信息", 
                "description": "航班、火车、汽车等交通工具信息"
            },
            {
                "id": "accommodation",
                "name": "住宿信息",
                "description": "酒店、民宿等住宿选择"
            },
            {
                "id": "activities",
                "name": "活动体验",
                "description": "各种旅游活动和体验项目"
            },
            {
                "id": "practical",
                "name": "实用信息",
                "description": "签证、货币、天气、安全等实用信息"
            }
        ]
        
        for cat_data in default_categories:
            category = KnowledgeCategory(**cat_data)
            self.categories[category.id] = category
    
    async def _load_knowledge_data(self):
        """加载知识数据"""
        # TODO: 从多个数据源加载知识
        # 1. 本地JSON/YAML文件
        # 2. 在线API数据
        # 3. 爬虫数据
        await self._load_local_knowledge()
        await self._load_default_travel_knowledge()
    
    async def _load_local_knowledge(self):
        """加载本地知识文件"""
        knowledge_files = self.knowledge_dir.glob("**/*.json")
        for file_path in knowledge_files:
            # TODO: 解析并加载知识文件
            pass
    
    async def _load_default_travel_knowledge(self):
        """加载默认旅游知识"""
        # TODO: 创建基础的旅游知识数据
        default_knowledge = [
            {
                "id": "paris_eiffel_tower",
                "title": "巴黎埃菲尔铁塔",
                "content": "埃菲尔铁塔是法国巴黎的标志性建筑，高324米，建于1889年。开放时间：9:00-23:00（夏季延长至24:00）。门票价格：成人29.4欧元（顶层），18.1欧元（二层）。",
                "category": "destinations",
                "location": "Paris, France",
                "tags": ["巴黎", "埃菲尔铁塔", "景点", "地标"]
            },
            {
                "id": "japan_visa_requirements",
                "title": "日本签证申请要求",
                "content": "中国公民赴日旅游需要申请短期滞在签证。申请材料包括：护照、签证申请表、照片、在职证明、银行流水、行程单等。处理时间通常为5-7个工作日。",
                "category": "practical",
                "location": "Japan",
                "tags": ["日本", "签证", "申请", "材料", "旅游"]
            }
        ]
        
        for knowledge_data in default_knowledge:
            knowledge = TravelKnowledge(**knowledge_data)
            self.knowledge_items[knowledge.id] = knowledge
    
    async def _build_index(self):
        """构建知识索引"""
        # TODO: 将知识数据索引到RAG引擎
        documents = []
        for knowledge in self.knowledge_items.values():
            doc = Document(
                id=knowledge.id,
                content=f"{knowledge.title}\n{knowledge.content}",
                metadata={
                    "category": knowledge.category,
                    "location": knowledge.location,
                    "tags": knowledge.tags,
                    "language": knowledge.language
                }
            )
            documents.append(doc)
        
        if documents:
            await self.rag_engine.index_documents(documents)
    
    async def add_knowledge(self, knowledge: TravelKnowledge) -> bool:
        """添加新知识"""
        # TODO: 添加新知识项
        # 1. 验证知识质量
        # 2. 添加到内存
        # 3. 更新索引
        # 4. 持久化存储
        pass
    
    async def update_knowledge(self, knowledge_id: str, updated_knowledge: TravelKnowledge) -> bool:
        """更新知识"""
        # TODO: 更新现有知识
        pass
    
    async def delete_knowledge(self, knowledge_id: str) -> bool:
        """删除知识"""
        # TODO: 删除知识项
        pass
    
    async def search_knowledge(
        self, 
        query: str, 
        category: Optional[str] = None,
        location: Optional[str] = None,
        top_k: int = 5
    ) -> List[TravelKnowledge]:
        """搜索知识"""
        # TODO: 基于RAG引擎搜索知识
        filter_metadata = {}
        if category:
            filter_metadata["category"] = category
        if location:
            filter_metadata["location"] = location
        
        result = await self.rag_engine.retrieve(
            query=query,
            top_k=top_k,
            filter_metadata=filter_metadata
        )
        
        # 将检索结果转换为知识对象
        knowledge_results = []
        for doc in result.documents:
            if doc.id in self.knowledge_items:
                knowledge_results.append(self.knowledge_items[doc.id])
        
        return knowledge_results
    
    def get_categories(self) -> List[KnowledgeCategory]:
        """获取所有分类"""
        return list(self.categories.values())
    
    def get_knowledge_by_category(self, category_id: str) -> List[TravelKnowledge]:
        """按分类获取知识"""
        return [
            knowledge for knowledge in self.knowledge_items.values()
            if knowledge.category == category_id
        ]


# 全局知识库实例
knowledge_base: Optional[KnowledgeBase] = None


async def get_knowledge_base() -> KnowledgeBase:
    """获取知识库实例"""
    global knowledge_base
    if knowledge_base is None:
        knowledge_base = KnowledgeBase()
        await knowledge_base.initialize()
    return knowledge_base 