requirements.txt
# 基础Web框架
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
python-multipart>=0.0.6

# AI和LLM相关
openai>=1.3.0
# anthropic>=0.7.0  # 可选：如果需要使用Claude

# RAG和向量数据库
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Agent框架（可选，根据需要启用）
# langchain>=0.1.0
# langchain-community>=0.0.10
# langchain-openai>=0.0.5

# 数据库和ORM
sqlalchemy>=2.0.23
alembic>=1.12.0  # 数据库迁移

# HTTP客户端和工具
aiohttp>=3.9.1
requests>=2.31.0
httpx>=0.25.1

# 数据处理和工具
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0.0

# 安全和认证
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1

# 日志和监控
loguru>=0.7.0
# opentelemetry-api>=1.20.0  # 可选：分布式追踪
# opentelemetry-sdk>=1.20.0

# 测试框架
pytest>=7.4.3
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# 开发工具
typing-extensions>=4.5.0
email-validator>=2.0.0

# 可选：如果需要Redis缓存
# redis>=4.5.0
# aioredis>=2.0.0

# 可选：如果需要PostgreSQL
# psycopg2-binary>=2.9.0
# asyncpg>=0.28.0

# 可选：机器学习和NLP
# scikit-learn>=1.3.0
# spacy>=3.6.0
# transformers>=4.30.0

# 性能优化
# uvloop>=0.18.0  # Unix系统性能优化 